combined button and delete popup
