+++ src/pages/dashboards/admin/editPages/AddLevel.jsx
@@ -93,6 +93,7 @@ const AddLevel = () => {
         setLocation("/admin/levels");
       }
     } catch (error) {
+      setIsSaving(false)
       console.error("Error adding level:", error);
       setAlertMessage(`Error: ${error.response.data.message}`);
       setTimeout(() => {
